WIP DO Deserializer using integer -> Long

+++ org.eclipse.scout.rt.jackson/src/main/java/org/eclipse/scout/rt/jackson/dataobject/DoEntityDeserializer.java
@@ -276,7 +276,7 @@ public class DoEntityDeserializer extends StdDeserializer<IDoEntity> {
     }
     if (currentToken == JsonToken.VALUE_NUMBER_INT) {
       // deserialize floating point numbers as BigDecimal
-      return AttributeType.ofDoValue(TypeFactory.defaultInstance().constructType(BigDecimal.class));
+      return AttributeType.ofDoValue(TypeFactory.defaultInstance().constructType(Long.class));
     }
     // JSON scalar values are deserialized as raw object using default jackson typing
     return AttributeType.ofDoValue(TypeFactory.unknownType());
